authorization with gates and policies

+++ app/Http/Controllers/IdeaController.php
@@ -7,6 +7,7 @@ use App\Models\Idea;
 use Illuminate\Http\Request;
 use App\Http\Requests\IdeaRequest;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class IdeaController extends Controller
 {
@@ -46,6 +47,9 @@ class IdeaController extends Controller
      */
     public function show(Idea $idea)
     {
+
+        Gate::authorize('update', $idea);
+
         return view('ideas/show', ['idea' => $idea]);
     }
 
@@ -54,6 +58,9 @@ class IdeaController extends Controller
      */
     public function edit(Idea $idea)
     {
+
+        Gate::authorize('update', $idea);
+
         return view('ideas/edit', ['idea' => $idea]);
     }
 
@@ -62,6 +69,9 @@ class IdeaController extends Controller
      */
     public function update(IdeaRequest $request, Idea $idea)
     {
+
+        Gate::authorize('update', $idea);
+
         $idea->update([
             'description' => request('description')
         ]);
@@ -74,17 +84,10 @@ class IdeaController extends Controller
      */
     public function destroy(Idea $idea)
     {
+
+        Gate::authorize('update', $idea);
+
         $idea->delete();
-
-        return redirect('/ideas');
-    }
-
-    /**
-     * Remove all resources of this type from storage.
-     */
-    public function destroyAll()
-    {
-        //Idea::truncate();
 
         return redirect('/ideas');
     }
